+++ CHANGES.rst
@@ -1,3 +1,13 @@
+unreleased
+==========
+
+- Rename the view predicates to ``last_retry_attempt`` and
+  ``retryable_error``. See https://github.com/Pylons/pyramid_retry/pull/3
+
+- Rename ``pyramid_retry.is_exc_retryable`` to
+  ``pyramid_retry.is_error_retryable``.
+  See https://github.com/Pylons/pyramid_retry/pull/3
+
 0.2 (2017-03-02)
 ================